+++ src/context/contextalbumsmodel.cpp
@@ -267,7 +267,7 @@ void ContextAlbumsModel::Reset() {
 
 CollectionItem *ContextAlbumsModel::ItemFromSong(CollectionItem::Type item_type, const bool signal, CollectionItem *parent, const Song &s, const int container_level) {
 
-  if (signal) beginInsertRows(ItemToIndex(parent), parent->children.count(), parent->children.count());
+  if (signal) beginInsertRows(ItemToIndex(parent), static_cast<int>(parent->children.count()), static_cast<int>(parent->children.count()));
 
   CollectionItem *item = new CollectionItem(item_type, parent);
   item->container_level = container_level;
