+++ src/context/contextview.cpp
@@ -715,9 +715,9 @@ void ContextView::ResetSong() {
 
 }
 
-void ContextView::UpdateLyrics(const quint64 id, const QString &provider, const QString &lyrics) {
+void ContextView::UpdateLyrics(const int id, const QString &provider, const QString &lyrics) {
 
-  if (static_cast<qint64>(id) != lyrics_id_) return;
+  if (id != lyrics_id_) return;
   lyrics_ = lyrics + "\n\n(Lyrics from " + provider + ")\n";
   lyrics_id_ = -1;
   if (action_show_lyrics_->isChecked()) {
