+++ src/core/utilities.h
@@ -159,13 +159,13 @@ class ScopedWCharArray {
   wchar_t *get() const { return data_.get(); }
   explicit operator wchar_t*() const { return get(); }
 
-  int characters() const { return chars_; }
-  int bytes() const { return (chars_ + 1)  *sizeof(wchar_t); }
+  qint64 characters() const { return chars_; }
+  qint64 bytes() const { return (chars_ + 1)  *sizeof(wchar_t); }
 
  private:
   Q_DISABLE_COPY(ScopedWCharArray)
 
-  int chars_;
+  qint64 chars_;
   std::unique_ptr<wchar_t[]> data_;
 };
